Refactor logging and metrics funcs out of loop.py
