+++ workflows/prognostic_c48_run/runtime/__init__.py
@@ -1,5 +1,4 @@
 from .config import get_namelist, get_config, write_chunks
-from .capture import capture_stream, capture_stream_mpi, capture_fv3gfs_funcs
 from .diagnostics import (
     get_chunks,
     get_diagnostic_files,
@@ -12,6 +11,15 @@ from .diagnostics import (
     default_diagnostics,
 )
 from .derived_state import DerivedFV3State
+from .logs import (
+    capture_stream,
+    capture_stream_mpi,
+    capture_fv3gfs_funcs,
+    setup_loggers,
+    log_scalar,
+    log_profiles,
+)
+from .metrics import globally_average_2d_diagnostics, globally_sum_3d_diagnostics
 from .nudging import (
     nudging_timescales_from_dict,
     setup_get_reference_state,
